client -> tip of the chain: extract new block handler addition to dedicated method, expand to light sync
