+++ packages/client/lib/sync/lightsync.ts
@@ -74,43 +74,49 @@ export class LightSynchronizer extends Synchronizer {
    * @return Resolves when sync completed
    */
   async syncWithPeer(peer?: Peer): Promise<boolean> {
-    if (!peer) return false
-    const height = new BN(peer.les!.status.headNum)
-    const first = this.chain.headers.height.addn(1)
-    const count = height.sub(first).addn(1)
-    if (count.lten(0)) return false
+    // eslint-disable-next-line no-async-promise-executor
+    return await new Promise(async () => {
+      if (!peer) return false
 
-    this.config.logger.debug(
-      `Syncing with peer: ${peer.toString(true)} height=${height.toString(10)}`
-    )
+      const latest = await this.latest(peer)
+      if (!latest) return false
+      const height = new BN(peer.les!.status.headNum)
+      const first = this.chain.headers.height.addn(1)
+      const count = height.sub(first).addn(1)
+      if (count.lten(0)) return false
 
-    this.headerFetcher = new HeaderFetcher({
-      config: this.config,
-      pool: this.pool,
-      chain: this.chain,
-      flow: this.flow,
-      interval: this.interval,
-      first,
-      count,
-    })
-    this.config.events.on(Event.SYNC_FETCHER_FETCHED, (headers) => {
-      headers = headers as BlockHeader[]
-      const first = new BN(headers[0].number)
-      const hash = short(headers[0].hash())
-      const baseFeeAdd = this.config.chainCommon.gteHardfork('london')
-        ? `basefee=${headers[0].baseFeePerGas} `
-        : ''
-      this.config.logger.info(
-        `Imported headers count=${headers.length} number=${first.toString(
-          10
-        )} hash=${hash} ${baseFeeAdd}peers=${this.pool.size}`
+      this.config.logger.debug(
+        `Syncing with peer: ${peer.toString(true)} height=${height.toString(10)}`
       )
+
+      this.headerFetcher = new HeaderFetcher({
+        config: this.config,
+        pool: this.pool,
+        chain: this.chain,
+        flow: this.flow,
+        interval: this.interval,
+        first,
+        count,
+        destroyWhenDone: false,
+      })
+      const headerFetcher = <HeaderFetcher>this.headerFetcher
+      this.addNewBlockHandlers(peer, headerFetcher)
+
+      this.config.events.on(Event.SYNC_FETCHER_FETCHED, (headers) => {
+        headers = headers as BlockHeader[]
+        const first = new BN(headers[0].number)
+        const hash = short(headers[0].hash())
+        const baseFeeAdd = this.config.chainCommon.gteHardfork('london')
+          ? `basefee=${headers[0].baseFeePerGas} `
+          : ''
+        this.config.logger.info(
+          `Imported headers count=${headers.length} number=${first.toString(
+            10
+          )} hash=${hash} ${baseFeeAdd}peers=${this.pool.size}`
+        )
+      })
+      await this.headerFetcher.fetch()
     })
-    await this.headerFetcher.fetch()
-    // TODO: Should this be deleted?
-    // @ts-ignore: error: The operand of a 'delete' operator must be optional
-    delete this.headerFetcher
-    return true
   }
 
   /**
